requirements.txt
Flask==2.2.2
names==0.3.0
prettytable==3.6.0
PyMySQL==1.0.2
PyPika==0.48.9
requests==2.28.2
selenium==4.8.0
cryptography==39.0.1
python-dotenv~=0.21.1
load-dotenv~=0.1.0
docker~=6.0.1
psutil~=5.9.4
async-generator==1.10
attrs==22.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
outcome==1.2.0
pip==23.0
pycparser==2.21
PySocks==1.7.1
setuptools==67.2.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wcwidth==0.2.6
Werkzeug==2.2.2
wheel==0.38.4
wsproto==1.2.0
kubernetes~=26.1.0
ruamel.yaml~=0.17.21
